Database connectivity and spotify features
Connection to a mongoDB database has been made, and is handled by a new
javascript file on the server. New features have been added, such as
tokens being served from the client (which should help with the multple
users thing) and i've figured out how to play back songs at a certain
timestamp, and start a user's song at a given timestamp, which should
provide the "live" song sync. Not sure how we are going to sort songs
yet, a few functions will have to be changed to handle the fact songs
will change in priority