Radio is now working
Songs can now be chosen from the searc bar, and will apear in the list.
the list is ordered by votes and timing of choice, but the buttons to
vote need to be implemented. Songs will then be played love on the
radio, and clients will be synced to the radio as it is playing. a
client can relog and should be able to relocate themself on the radio.
just need front-end UI changes and an API call to apply vote changes.
Maybe a refresh button as well :smile: :chart_with_upwards_trend:
:hamburger: