Cleaned and finished
Design overhaul to give the project the final polish. Test functionality
added to show how the program works in the backgroung, incliding the
button to get a new refresh token. voting works (if you don't spam the
buttons, for some reason this can jam the system and only one song will
receive the votes) :checkered_flag: